requirements.txt
google-adk>=1.0.0
google-cloud-bigquery>=3.10.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-httplib2>=0.2.0
google-cloud-logging>=3.5.0
google-cloud-aiplatform[reasoningengine]>=1.140.0
google-cloud-storage>=2.10.0
google-cloud-discoveryengine>=0.11.0
python-dotenv>=1.0.0
pydantic>=2.0.0
deprecated>=1.2.14
pytest>=7.0.0
pytest-cov>=4.0.0
